New Files
I have added in three new files for the program.
First is the includes file. This will share common imports to all files so we only need to declare things once.
Second/Third is the the new VehicleComputer. This is going to be used to control a Vehicle, this means how it drives and being able to calculate its information. Later this will hold AI for the cars to drive and will be able to read in the environment and react based on it.